Custom Scripts/WSA-Migration-AN/PRC/requirements.txt
Flask==2.0.2
brotli==1.0.9
brotlicffi==1.0.9.2
ConfigParser==5.3.0
cryptography==36.0.1
Cython==0.29.33
dl==0.1.0
docutils==0.19
HTMLParser==0.0.2
importlib_metadata==6.0.0
keyring==23.13.1
lockfile==0.12.2
lxml==4.7.1
mock==5.0.1
numpy==1.22.1
ordereddict==1.1
pandas==1.4.1
protobuf==4.21.12
pyOpenSSL==23.0.0
simplejson==3.18.1
toml==0.10.2
tornado==6.2
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
zeep==4.1.0
zipp==3.12.0
